+++ src/utility.cpp
@@ -7,6 +7,7 @@
 #include "utility.hpp"
 #include "platform.hpp"
 #include "rpgmakerVersion.hpp"
+#include "foldertype.hpp"
 
 bool isValidDirectory(const std::string& directory, const std::string& name, const std::shared_ptr<spdlog::logger>& errorLogger) {
     if (!ghc::filesystem::exists(directory)) {
@@ -328,4 +329,86 @@ RPGMakerVersion getRPGMakerVersion(const ghc::filesystem::path& projectPath, con
     }
 
     return version;
+}
+
+bool filterFile(ghc::filesystem::path* from, ghc::filesystem::path* to, FolderType folderType, RPGMakerVersion version, Platform platform) {
+    auto filename = from->filename();
+    auto extension = from->extension();
+
+    auto parent = from->parent_path();
+    auto parentName = parent.filename();
+
+    if (folderType == FolderType::RPGMaker) {
+        if (version == RPGMakerVersion::MZ) {
+            if (parentName == "pnacl") return true;
+
+            //nw.exe gets renamed to Game.exe to reduce confusion for players
+            if (filename == "nw.exe") {
+                to->replace_filename("Game.exe");
+                return false;
+            }
+
+            //chromium related files that are ignored
+            //TODO: Windows only?
+            if (filename == "chromedriver.exe") return true;
+            if (filename == "nacl_irt_x86_64.nexe") return true;
+            if (filename == "nw.exe") return true;
+            if (filename == "nwjc.exe") return true;
+            if (filename == "payload.exe") return true;
+        }
+    } else if (folderType == FolderType::Project) {
+        //Desktop: only ogg
+        //Mobile: only m4a
+        //Browser: both
+        if (platform == Platform::Mobile && extension == ".ogg") return true;
+        if (extension == ".m4a" && platform != Platform::Mobile && platform != Platform::Browser) return true;
+
+        //skip project files
+        if (extension == ".rpgproject") return true;
+        if (extension == ".rmmzproject") return true;
+    }
+
+    return false;
+}
+
+bool shouldEncryptFile(ghc::filesystem::path *from, bool encryptAudio, bool encryptImages, RPGMakerVersion version) {
+    auto filename = from->filename();
+    auto extension = from->extension();
+    auto parent = from->parent_path();
+    auto parentName = parent.filename();
+    auto grandparent = parent.parent_path();
+    auto grandparentName = grandparent.filename();
+
+    if (extension != ".png" && extension != ".ogg" && extension != ".m4a")
+        return false;
+
+    if (encryptAudio && (extension == ".ogg" || extension == ".m4a")) {
+        return true;
+    }
+
+    if (encryptImages && extension == ".png") {
+        if (version == RPGMakerVersion::MZ) {
+            //effects/Textures is not encrypted in MZ
+            if (parentName == "Texture" && grandparentName == "effects")
+                return false;
+
+            //img/system gets encrypted, in MV some are left out
+            if (parentName == "system" && grandparentName == "img")
+                return true;
+        } else if (version == RPGMakerVersion::MV) {
+            if (parentName == "system" && grandparentName == "img") {
+                //these are for some reason not encrypted in MV
+                if (filename == "Loading.png") return false;
+                if (filename == "Window.png") return false;
+            }
+        }
+
+        //game icon for the window
+        if (filename == "icon.png" && parentName == "icon")
+            return false;
+
+        return true;
+    }
+
+    return false;
 }
